refactor: session 패키지 변경
