+++ tomcat/src/main/java/org/apache/catalina/manager/SessionManager.java
@@ -1,6 +1,6 @@
 package org.apache.catalina.manager;
 
-import org.apache.catalina.Session;
+import org.apache.coyote.Session;
 
 import java.util.HashMap;
 import java.util.Map;
